dashboard/requirements.txt
streamlit
google-cloud-firestore
google-auth
pandas
requests
altair
python-dotenv